pyproject.toml
[project]
name = "prompt-extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "langchain>=0.3.27",
    "langchain-core>=0.3.81",
    "pdfplumber>=0.11.9",
    "pydantic>=2.12.5",
    "pypdf>=6.5.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "sap-ai-sdk-gen>=5.10.0",
    "uvicorn[standard]>=0.40.0",
    "yaspin>=3.4.0",
]
